implemented register_shutdown_function support
It is often the case that you do not need the ES index updated inline,
and could (wait) until after the page has been returned to the requester.

In this case, you'd just set `register_shutdown_function = true`
and the Behavior doesn't attempt to write to ElasticSearch until after
the script is complete (the page returned to the user).

NOTE: PHP-FPM users, you are going to have to call `fastcgi_finish_request`
because exit() doesn't terminate php-fpm calls immediately.

You can use the following at the very END of your scripts (if you have to)

    /* ready for register_shutdown_function() to process? */
    if (function_exists('fastcgi_finish_request')) {
      /* force a FPM finish request (like an exit) */
      fastcgi_finish_request();
    }
    /* force an exit */
    exit;